requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
redis==4.6.0
celery==5.3.4
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
tenacity==8.2.3
python-dateutil==2.8.2
